src/tool/custom_tools/browser_use_with_cua/requirements.txt
browser-use
openai>=1.0.0
python-dotenv
fastmcp
annotated-types==0.7.0
anyio>=4.9.0,<5
browserbase==1.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
distro==1.9.0
fastapi==0.116.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.8.2
MarkupSafe==3.0.2
pillow==11.1.0
playwright==1.50.0
pydantic>=2.11.7
pyee==12.1.1
python-multipart==0.0.20
requests==2.32.3
scrapybara==2.5.3
sniffio==1.3.1
starlette==0.47.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.35.0
mcp==1.13.1